Various updates to the sim code: * Use half precision floats for height texture * Add methods to sample simulation data * Allow solid data to be pre-initialized.
And the demo scene has been switched over to periodic inlet/outlet.

+++ com.kircher.lbmswe/Runtime/UpdateHeightTextureJob.cs
@@ -15,9 +15,9 @@ namespace LatticeBoltzmannMethods
         [ReadOnly]
         private NativeArray<float> _waterHeight;
         [NativeDisableParallelForRestriction]
-        private NativeArray<float> _destination;
+        private NativeArray<half> _destination;
 
-        public UpdateHeightTextureJob(int textureWidth, float maxHeight, NativeArray<float> waterHeight, NativeArray<float> destination)
+        public UpdateHeightTextureJob(int textureWidth, float maxHeight, NativeArray<float> waterHeight, NativeArray<half> destination)
         {
             _textureWidth = textureWidth;
             _maxHeight = maxHeight;
@@ -31,7 +31,7 @@ namespace LatticeBoltzmannMethods
             for (var idx = startIdx; idx < startIdx + _textureWidth; ++idx)
             {
                 var height = _waterHeight[idx];
-                _destination[idx] = math.saturate(height / _maxHeight);
+                _destination[idx] = (half)math.saturate(height / _maxHeight);
             }
         }
     }
